+++ alerter/src/alerter/metric_code/internal_metric_code.py
@@ -2,5 +2,4 @@ from src.alerter.metric_code.metric_code import MetricCode
 
 
 class InternalMetricCode(MetricCode):
-    ComponentResetChains = 'component_reset'
-    ComponentResetAllChains = 'component_reset_all'
+    ComponentResetAlert = 'component_reset_alert'